Build: add VS Code headless release workflow and clean warning/linker fixes
- document STM32CubeIDE headless CLI build flow in README
- add note that bootloader build is not hardware-tested yet
- ignore build-logs and .DS_Store in root .gitignore
- set Release optimization to O2 in project configs
- fix clean/build order and logging flow in scripts/build_cli.sh
- fix objcopy conversion to exclude RAM sections from .bin output
- resolve firmware/lib/bootloader warnings for GCC 14/15 compatibility
- remove RWX LOAD segment linker warnings via linker script PHDR updates
- keep release clean build at 0 errors / 0 warnings
- note: VS Code/CLI workflow and warning-cleanup changes

+++ firmware/Src/utils/Hexter.cpp
@@ -309,7 +309,7 @@ float Hexter::getRounded(float r) {
  */
 void Hexter::voiceSetData(struct OneSynthParams *params, uint8_t *patch)
 {
-    for (int k=0; k<sizeof(struct OneSynthParams)/sizeof(float); k++) {
+	for (int k = 0; k < (int)(sizeof(struct OneSynthParams) / sizeof(float)); k++) {
         int row = k >> 2;
         if (row == ROW_ARPEGGIATOR1 || row == ROW_ARPEGGIATOR2 || row == ROW_ENGINE) {
             // Don't override the arpeggiator neither the ENGINE
